Fix wording around "Authenticate in admin"
The original wording was unclear to me. After it was explained in the comments for [issue 2484](https://github.com/ioBroker/ioBroker.admin/issues/2484), I believe this change best reflects what was originally intended. This should make it easier for users to understand what they are being asked to decide upon.