Fix wording around "Authenticate in admin" (#2485)
The original wording was unclear to me. After it was explained in the
comments for [issue
2484](https://github.com/ioBroker/ioBroker.admin/issues/2484), I believe
this change best reflects what was originally intended. This should make
it easier for users to understand what they are being asked to decide
upon.

I only updated the English translation in this edit. In my opinion the
`de` and `nl` versions are similarly confusing and need updates too. I
suspect this also applies to other languages, but I haven't checked.

I can provide these changes for `nl` once this change is accepted.

Co-authored-by: Bluefox <dogafox@gmail.com>